[util] move limitlistener to util for other protocol proxy use
