+++ pkg/object/httpserver/runtime.go
@@ -31,6 +31,7 @@ import (
 	"github.com/megaease/easegress/pkg/protocol"
 	"github.com/megaease/easegress/pkg/supervisor"
 	"github.com/megaease/easegress/pkg/util/httpstat"
+	"github.com/megaease/easegress/pkg/util/limitlistener"
 	"github.com/megaease/easegress/pkg/util/topn"
 )
 
@@ -81,7 +82,7 @@ type (
 
 		httpStat      *httpstat.HTTPStat
 		topN          *topn.TopN
-		limitListener *LimitListener
+		limitListener *limitlistener.LimitListener
 	}
 
 	// Status contains all status generated by runtime, for displaying to users.
@@ -275,7 +276,7 @@ func (r *runtime) startServer() {
 			return
 		}
 
-		limitListener := NewLimitListener(listener, r.spec.MaxConnections)
+		limitListener := limitlistener.NewLimitListener(listener, r.spec.MaxConnections)
 		r.limitListener = limitListener
 		go r.runHTTP1And2Server(limitListener, r.spec.HTTPS, r.startNum)
 	}
@@ -291,7 +292,7 @@ func (r *runtime) runHTTP3Server(startNum uint64) {
 	}
 }
 
-func (r *runtime) runHTTP1And2Server(limitListener *LimitListener, https bool, startNum uint64) {
+func (r *runtime) runHTTP1And2Server(limitListener *limitlistener.LimitListener, https bool, startNum uint64) {
 	var err error
 	if https {
 		err = r.server.ServeTLS(limitListener, "", "")
