+++ pkg/util/limitlistener/limitlistener.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package httpserver
+package limitlistener
 
 import (
 	"context"
